Add 2-decimal precision to Powerlevel score
Score and powerlevel now round to 2 decimals instead of whole
numbers. Continuous factors (tutors, fast mana, mana base) keep
fractional output; threshold-based factors stay integer per spec.
Powerlevel column changed to Float to persist decimal values.

+++ src/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, Float, String, ForeignKey
 from sqlalchemy.orm import relationship
 from .database import Base
 
@@ -21,7 +21,7 @@ class Deck(Base):
     image_url = Column(String)
     archetype = Column(String, nullable=True)
     bracket = Column(String, nullable=True)
-    powerlevel = Column(Integer, nullable=True)
+    powerlevel = Column(Float, nullable=True)
     status = Column(String)
     moxfield_link = Column(String, nullable=True)
